ai/requirements.txt
pillow
scikit-learn
tqdm
ultralytics
